requirements.txt
matplotlib==3.9.2
numpy==2.1.3
Pillow==11.0.0
streamlit==1.34.0
